pyproject.toml
[project]
name = "lab-hello-world-2nd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "async>=0.6.2",
    "fastapi>=0.116.0",
    "git-ignore>=1.0.2",
    "httpx>=0.28.1",
    "pyyaml>=6.0.2",
    "streamlit>=1.46.1",
    "uvicorn>=0.35.0",
]

[tool.uv.workspace]
members = [
    "git.ignore",
]
